+++ sparkplug/src/drivers/sn3218driver.cpp
@@ -25,19 +25,36 @@ void SN3218Driver::setup()
 
     for (int i = 0; i < outputCount; i++)
         set(i, 0);
+
+    sn3218.update();
 };
 
 void SN3218Driver::output()
 {
-    for (int i = 0; i < channelCount; i++)
+    bool lastChannelUpdated = false;
+
+    for (int i = 0; i <= channelCount; i++)
     {
-        if (!channels[channelFrom + i].wasUpdated) continue;
+        if (channels[channelFrom + i].wasUpdated)
+        {
+            if (!lastChannelUpdated)
+            {
+                Wire.beginTransmission(SN3218_ADDR);
+                Wire.write(SN3218_CMD_SET_PWM_VALUES + i);
+            }
 
-        uint8_t value = channels[channelFrom + i].value >> 8;
-        uint8_t logValue = PROGMEM_getAnything(&gamma8[value]);
-
-        set(i, logValue);
+            uint8_t value = channels[channelFrom + i].value >> 8;
+            uint8_t logValue = PROGMEM_getAnything(&gamma8[value]);
+            Wire.write(logValue);
+            lastChannelUpdated = true;
+        }
+        else
+        {
+            if (lastChannelUpdated) Wire.endTransmission();
+            lastChannelUpdated = false;
+        }
     }
+    if (lastChannelUpdated) Wire.endTransmission();
 
     sn3218.update();
 };